Updated integers to unsignedInteger() in migrations to match unique AI ids

+++ database/migrations/2017_09_14_082032_create_torrents_table.php
@@ -18,10 +18,10 @@ class CreateTorrentsTable extends Migration
             $table->string('info_hash', 40)->unique();
             $table->string('name', 500);
             $table->text('description')->default('');
-            $table->bigInteger('category')->default(1);
+            $table->unsignedInteger('category')->default(1);
             $table->bigInteger('size')->default(0);
             $table->bigInteger('downloads')->default(0);
-            $table->bigInteger('user')->default(0);
+            $table->unsignedInteger('user')->default(0);
 
             $table->timestamps();
         });
