+++ database/migrations/2017_09_14_221031_create_files_table.php
@@ -15,7 +15,7 @@ class CreateFilesTable extends Migration
     {
         Schema::create('files', function (Blueprint $table) {
             $table->increments('id');
-            $table->bigInteger('torrent');
+            $table->unsignedInteger('torrent');
             $table->string('path');
             $table->bigInteger('size');
